[PyTorch] Make c10::str(const char*) return const char*
`c10::str()` is often used with variadic macros. It can be more efficient to get a C string out if you put a C string in, like if you are able to defer std::string creation to an outlined function or even never do it at all. Meanwhile, there is an implicit conversion from const char* to std::string, so users who expected a std::string will still make one.

Differential Revision: [D26419663](https://our.internmc.facebook.com/intern/diff/D26419663/)

ghstack-source-id: 121619713
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52222